store/requirements.txt
django==1.11
djangorestframework==3.6.2
psycopg2==2.7.3.2